requirements.txt
blinker==1.9.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
ftfy==6.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.1
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
